--- src/organisation/containers/index.ts
@@ -1,5 +1,0 @@
-import { OrganisationComponent } from './organisation/organisation.component';
-
-export const containers: any[] = [OrganisationComponent];
-
-export * from './organisation/organisation.component';
