--- src/organisation/containers/organisation/organisation.component.html
@@ -1,60 +1,0 @@
-<app-hmcts-main-wrapper [title]="'Organisation'">
-  <table class="govuk-table">
-    <tbody class="govuk-table__body">
-    <tr class="govuk-table__row">
-      <th class="govuk-table__header" scope="row">Name</th>
-      <td class="govuk-table__cell">
-        <div>{{organisationDetails.name}}</div>
-      </td>
-    </tr>
-    <tr class="govuk-table__row">
-      <th class="govuk-table__header" scope="row">Address</th>
-      <td class="govuk-table__cell">
-        <!-- Contact Information -->
-        <div>{{organisationContactInformation.addressLine1}}</div>
-        <div *ngIf="organisationContactInformation.addressLine2">{{organisationContactInformation.addressLine2}}
-        </div>
-        <div>{{organisationContactInformation.townCity}}</div>
-        <div>{{organisationContactInformation.county}}</div>
-        <div *ngIf="organisationContactInformation.postCode">{{organisationContactInformation.postCode}}
-        </div>
-        <!-- !Contact Information -->
-      </td>
-    </tr>
-    <!-- PBA Numbers -->
-    <tr class="govuk-table__row" *ngIf="organisationPaymentAccount">
-      <th class="govuk-table__header" scope="row">PBA number</th>
-      <td class="govuk-table__cell">
-        <div *ngFor="let pbaNumber of organisationPaymentAccount">{{pbaNumber}}</div>
-      </td>
-    </tr>
-    <!-- !PBA Numbers -->
-    <tr class="govuk-table__row" *ngIf="organisationDxAddress && organisationDxAddress.dxNumber">
-      <th class="govuk-table__header" scope="row">DX number</th>
-      <td class="govuk-table__cell">
-        <div>{{organisationDxAddress.dxNumber}}</div>
-      </td>
-    </tr>
-    <tr class="govuk-table__row" *ngIf="organisationDxAddress && organisationDxAddress.dxExchange">
-      <th class="govuk-table__header" scope="row">DX exchange</th>
-      <td class="govuk-table__cell">
-        <div>{{organisationDxAddress.dxExchange}}</div>
-      </td>
-    </tr>
-    <tr class="govuk-table__row" *ngIf="organisationDetails.superUser">
-      <th class="govuk-table__header" scope="row">Administrator</th>
-      <td class="govuk-table__cell">
-        <div>{{organisationDetails.superUser.firstName}} {{organisationDetails.superUser.lastName}}</div>
-        <div class="govuk-caption-m govuk-!-font-size-16">{{organisationDetails.superUser.email}}</div>
-      </td>
-    </tr>
-    <tr class="govuk-table__row" *ngIf="organisationDetails.sraId">
-      <th class="govuk-table__header" scope="row">SRA ID</th>
-      <td class="govuk-table__cell">
-        <div>{{organisationDetails.sraId}}</div>
-      </td>
-    </tr>
-    </tbody>
-  </table>
-</app-hmcts-main-wrapper>
-
