--- src/organisation/organisation.routing.ts
@@ -1,18 +1,0 @@
-// routes
-import { RouterModule, Routes } from '@angular/router';
-import { ModuleWithProviders } from '@angular/core';
-import { OrganisationComponent } from './containers';
-import { OrganisationGuard } from 'src/organisation/guards/organisation.guard';
-
-export const ROUTES: Routes = [
-  {
-    path: '',
-    component: OrganisationComponent,
-    canActivate: [
-      OrganisationGuard
-    ]
-  }
-];
-
-
-export const organisationRouting: ModuleWithProviders = RouterModule.forChild(ROUTES);
