+++ src/organisation/store/reducers/organisation.reducer.spec.ts
@@ -1,6 +1,6 @@
 import * as fromOrganisation from './organisation.reducer';
 import * as fromActions from '../actions/organisation.actions';
-import { Organisation } from 'src/organisation/organisation.model';
+import { Organisation } from 'src/organisation/models/organisation.interface';
 
 
 describe('OrganisationReducer', () => {
