+++ src/organisation/store/selectors/organisation.selectors.spec.ts
@@ -4,7 +4,7 @@ import { OrganisationState } from '../reducers/organisation.reducer';
 import { getOrganisationSel, getOrganisationState } from './organisation.selectors';
 import { reducers } from '../index';
 import { LoadOrganisationSuccess } from '../actions';
-import { Organisation } from 'src/organisation/organisation.model';
+import { Organisation } from 'src/organisation/models/organisation.interface';
